Add the ported module with accompanying license files.
Porting of the code from Python to PHP was completed. There shouldn't be
any bugs. *crosses fingers*

PHP's getopt() function totally sucks, so I've had to port Python's
gnu_getopt() function to PHP as well because trying to make sense out of
PHP's madness and to adapt that to fit into the original code was
making me go crazy.
So, sorry folks, the file is 500+ lines long instead of 300-ish.